requirements.txt
streamlit
pyyaml
huggingface-hub